1.7.11

+++ game/background.js
@@ -57,10 +57,18 @@
     }
     window.resetGameTimeout=setTimeout(function(){
         if(window.inSplash) return;
-        if(confirm('游戏似乎未正常载入，是否重置游戏？')){
-            localStorage.clear();
-            if(indexedDB) indexedDB.deleteDatabase('noname_0.9_data');
-            window.location.reload();
+        if(window.resetExtension){
+            if(confirm('游戏似乎未正常载入，是否禁用扩展并重新打开？')){
+                window.resetExtension();
+                window.location.reload();
+            }
+        }
+        else{
+            if(confirm('游戏似乎未正常载入，是否重置游戏？')){
+                localStorage.clear();
+                if(indexedDB) indexedDB.deleteDatabase('noname_0.9_data');
+                window.location.reload();
+            }
         }
     },5000);
 }())
